+++ firmware/src/ui.rs
@@ -5,6 +5,7 @@ use crate::error::Error;
 use crate::http::Http;
 use crate::keypad::{Key, Keypad};
 use crate::nfc::Nfc;
+use crate::schedule::Daily;
 use crate::screen;
 use crate::user::{UserId, Users};
 use crate::vereinsflieger::Vereinsflieger;
@@ -49,6 +50,7 @@ pub struct Ui<'a, RNG, I2C, IRQ> {
     vereinsflieger: &'a mut Vereinsflieger<'a>,
     articles: &'a mut Articles<1>,
     users: &'a mut Users,
+    schedule: &'a mut Daily,
 }
 
 impl<'a, RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'a, RNG, I2C, IRQ> {
@@ -65,6 +67,7 @@ impl<'a, RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'a, RNG, I2C,
         vereinsflieger: &'a mut Vereinsflieger<'a>,
         articles: &'a mut Articles<1>,
         users: &'a mut Users,
+        schedule: &'a mut Daily,
     ) -> Self {
         Self {
             rng,
@@ -77,6 +80,7 @@ impl<'a, RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'a, RNG, I2C,
             vereinsflieger,
             articles,
             users,
+            schedule,
         }
     }
 
@@ -94,12 +98,8 @@ impl<'a, RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'a, RNG, I2C,
 
         self.display.screen(&screen::Splash).await?;
 
-        match with_timeout(SPLASH_TIMEOUT, self.keypad.read()).await {
-            // Key pressed
-            Ok(_key) => Ok(()),
-            // User interaction timeout
-            Err(TimeoutError) => Err(Error::UserTimeout),
-        }
+        let _ = with_timeout(SPLASH_TIMEOUT, self.keypad.read()).await;
+        Ok(())
     }
 
     /// Show error screen and wait for keypress or timeout
@@ -162,6 +162,8 @@ impl<'a, RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'a, RNG, I2C,
         self.display
             .screen(&screen::PleaseWait::UpdatingData)
             .await?;
+
+        // Connect to Vereinsflieger API
         let mut vf = self.vereinsflieger.connect(self.http).await?;
 
         // Show authenticated user information when debugging
@@ -179,6 +181,9 @@ impl<'a, RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'a, RNG, I2C,
 
     /// Initialize user interface
     pub async fn init(&mut self) -> Result<(), Error> {
+        // Show splash screen for a while
+        self.show_splash().await?;
+
         // Wait for network to become available (if not already)
         self.wait_network_up().await?;
 
@@ -190,13 +195,21 @@ impl<'a, RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'a, RNG, I2C,
 
     /// Run the user interface flow
     pub async fn run(&mut self) -> Result<(), Error> {
-        // Wait for id card and verify identification
-        let user_id = self.authenticate_user().await?;
+        // Either wait for id card read or schedule time
+        let schedule_timer = self.schedule.timer();
+        let user_id = match select(self.authenticate_user(), schedule_timer).await {
+            // Id card read
+            Either::First(res) => res?,
+            // Schedule time
+            Either::Second(()) => return self.schedule().await,
+        };
+
+        // Get user information
         let user = self.users.get(user_id);
+        let user_name = user.map_or(String::new(), |u| u.name.clone());
 
         // Ask for number of drinks
-        let name = user.map_or(String::new(), |u| u.name.clone());
-        let num_drinks = self.get_number_of_drinks(&name).await?;
+        let num_drinks = self.get_number_of_drinks(&user_name).await?;
 
         // Get article information
         let article_id = self.articles.id(0).ok_or(Error::ArticleNotFound)?.clone();
@@ -217,6 +230,20 @@ impl<'a, RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'a, RNG, I2C,
         // Show success and affirm to take drinks
         self.show_success(num_drinks).await?;
 
+        Ok(())
+    }
+
+    /// Run schedule
+    pub async fn schedule(&mut self) -> Result<(), Error> {
+        if self.schedule.is_expired() {
+            info!("UI: Running schedule...");
+
+            // Schedule next event
+            self.schedule.schedule_next();
+
+            // Refresh article and user information
+            self.refresh_articles_and_users().await?;
+        }
         Ok(())
     }
 }
@@ -274,15 +301,15 @@ impl<RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'_, RNG, I2C, IRQ
             #[allow(clippy::match_same_arms)]
             match with_timeout(USER_TIMEOUT, self.keypad.read()).await {
                 // Any digit 1..=9 selects number of drinks
-                Ok(Key::Digit(n)) if (1..=9).contains(&n) => return Ok(n as usize),
+                Ok(Key::Digit(n)) if (1..=9).contains(&n) => break Ok(n as usize),
                 // Ignore any other digit
                 Ok(Key::Digit(_)) => (),
                 // Cancel key cancels
-                Ok(Key::Cancel) => return Err(Error::Cancel),
+                Ok(Key::Cancel) => break Err(Error::Cancel),
                 // Ignore any other key
                 Ok(_) => (),
                 // User interaction timeout
-                Err(TimeoutError) => return Err(Error::UserTimeout),
+                Err(TimeoutError) => break Err(Error::UserTimeout),
             }
         }
     }
@@ -300,13 +327,13 @@ impl<RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'_, RNG, I2C, IRQ
         loop {
             match with_timeout(USER_TIMEOUT, self.keypad.read()).await {
                 // Enter key confirms purchase
-                Ok(Key::Enter) => return Ok(()),
+                Ok(Key::Enter) => break Ok(()),
                 // Cancel key cancels
-                Ok(Key::Cancel) => return Err(Error::Cancel),
+                Ok(Key::Cancel) => break Err(Error::Cancel),
                 // Ignore any other key
                 Ok(_) => (),
                 // User interaction timeout
-                Err(TimeoutError) => return Err(Error::UserTimeout),
+                Err(TimeoutError) => break Err(Error::UserTimeout),
             }
         }
     }
@@ -328,6 +355,8 @@ impl<RNG: RngCore, I2C: I2c, IRQ: Wait<Error = Infallible>> Ui<'_, RNG, I2C, IRQ
         );
 
         self.display.screen(&screen::PleaseWait::Purchasing).await?;
+
+        // Connect to Vereinsflieger API
         let mut vf = self.vereinsflieger.connect(self.http).await?;
 
         // Store purchase