+++ firmware/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Automatically refresh article and user information once a day
+
 ## 0.2.0 - 2024-11-27
 
 - Show random greetings to user